Update aPathfinding.js
Renamed some variables, got rid of some variables. Fixed a direction issue that caused the wrong direction to be returned. Renamed some parameters.
Fixed an issue that caused the direction to return the wrong direction (separate than the above issue). This was due to the angle not being positive. (This also caused flickering) and made the movement feel jerky.

Changed the distance at which you need to be from the tile you are moving to, to be considered in the node and trigger the move to the next node or end the pathfinder.