Update aPathfinding.js
Fixed an issue where the pathfinding system would never reach its target if it was a tile that had a yPos of 0. The distance checker would always fall short of where it needed to be to count as being on that node. This was due to an attempt at handling an engine quirk. Where negative positions of -10000 means the diob is not technically on the map. So the plugin clamped positions to be 0 if they became negative. This introduced the issue that I'm explaining. Diobs positions can be negative, and that is just fine. However if it is -10000 or above it is safe to say it should not be allowed to pathfind.