Update aPathfinding.js
Removed pause functionality.
Added a new event. DiobInstance.onPathStuck() // Called when the system notices you have been in the same position for too long.
Added a sanity check for aPathfindingWeight to make sure its a number before assigning it.
Added a pNearest parameter to DiobInstance.goTo This will attempt to move you to the closest endNode in the event that the end node is blocked.
pNearest will also try to find a better startPosition if you are currently blocked. (Not perfect)
Added more debug info to help later development.