add: default options for pathway system so no configuration needs to be passed.
add: min distance to add flexibility in controlling when a node is considered to be reached.

fix: `position` mode. This now moves the instance properly.

fix: max delta time is used when delta time reaches above 0.033

add: pixelsPerSecond to indicate the speed of the instance when using `position` mode.

fix: when determining the validity of a tile it did not take into account pathwayWeight. Now a tile is no longer marked invalid just because a dense instance is on it. Only when that dense instance has no pathwayWeight.